+++ django/applications/catmaid/static/js/init.js
@@ -965,7 +965,10 @@ var project;
     if (e.is_authenticated || (e.permissions && -1 !== e.permissions.indexOf('catmaid.can_browse'))) {
       // Asynchronously, try to get a full list of users if a user is logged in
       // or the anonymous user has can_browse permissions.
-      load = load.then(CATMAID.User.getUsers.bind(CATMAID.User));
+      load = load.then(Promise.all([
+        CATMAID.User.getUsers.bind(CATMAID.User),
+        CATMAID.Group.updateGroupCache(),
+      ]));
     }
 
     return load.then(done);